app/requirements.txt
tensorflow
numpy
pillow
tk
